Move `tokenrequestor` package from `pkg/resourcemanager/controller` to `pkg/controller`
It will be reused by `gardenlet` in the next commits, so let's move it to a neutral place.
